Move createScripts to server/render/
