+++ server/render/createScripts.js
